starknet_transaction_prover: extract graceful shutdown to shutdown.rs
Force-exit on a second termination signal now only applies while the drain is
still in progress; once the server has stopped cleanly a late signal no longer
flips the exit code to 1.
