+++ crates/starknet_transaction_prover/src/main.rs
@@ -25,14 +25,14 @@ async fn main() -> anyhow::Result<()> {
     use starknet_transaction_prover::server::panic::install_panic_hook;
     use starknet_transaction_prover::server::rpc_api::ProvingRpcServer;
     use starknet_transaction_prover::server::rpc_impl::ProvingRpcServerImpl;
+    use starknet_transaction_prover::server::shutdown::spawn_signal_bridge;
     use starknet_transaction_prover::server::{
         start_server,
         OhttpJsonrpseeLayer,
         OHTTP_JSONRPSEE_BODY_BUILDER,
     };
-    use tokio::signal::unix::{signal, Signal, SignalKind};
     use tower_ohttp::OhttpGateway;
-    use tracing::{info, warn};
+    use tracing::info;
     use tracing_subscriber::prelude::*;
     use tracing_subscriber::{fmt, EnvFilter};
 
@@ -48,9 +48,8 @@ async fn main() -> anyhow::Result<()> {
         LogFormat::Text => registry.with(fmt::layer()).init(),
     }
 
-    // Install after tracing init so the hook's `error!` macro reaches the
-    // subscriber. A panic before this line still hits the default stderr
-    // handler.
+    // After tracing init so the hook's `error!` reaches the subscriber;
+    // earlier panics hit the default stderr handler.
     install_panic_hook();
 
     let config = ServiceConfig::from_args(args)?;
@@ -115,65 +114,8 @@ async fn main() -> anyhow::Result<()> {
         "JSON-RPC proving server is running."
     );
 
-    // Waits for whichever termination signal fires first, returning its name
-    // (`None` if neither handler was installed). Awaiting the returned future
-    // borrows the handlers only for that wait, so it can be called again to
-    // block on a second signal.
-    async fn recv_termination_signal(
-        sigterm: &mut Option<Signal>,
-        sigint: &mut Option<Signal>,
-    ) -> Option<&'static str> {
-        match (sigterm, sigint) {
-            (Some(sigterm), Some(sigint)) => tokio::select! {
-                _ = sigterm.recv() => Some("SIGTERM"),
-                _ = sigint.recv() => Some("SIGINT"),
-            },
-            (Some(sigterm), None) => {
-                sigterm.recv().await;
-                Some("SIGTERM")
-            }
-            (None, Some(sigint)) => {
-                sigint.recv().await;
-                Some("SIGINT")
-            }
-            (None, None) => None,
-        }
-    }
-
-    // Bridge SIGTERM/SIGINT into jsonrpsee's `ServerHandle::stop` so
-    // container teardown becomes visible in logs. Both handlers are
-    // installed eagerly: if one fails, we still want the other to drive
-    // a graceful shutdown rather than silently dropping it.
-    let sigterm = signal(SignalKind::terminate())
-        .inspect_err(|err| warn!(error = %err, "Failed to install SIGTERM handler"))
-        .ok();
-    let sigint = signal(SignalKind::interrupt())
-        .inspect_err(|err| warn!(error = %err, "Failed to install SIGINT handler"))
-        .ok();
-    let shutdown_handle = server_handle.clone();
-    tokio::spawn(async move {
-        let (mut sigterm, mut sigint) = (sigterm, sigint);
-        let Some(signal_name) = recv_termination_signal(&mut sigterm, &mut sigint).await else {
-            return;
-        };
-        info!(event = "shutdown_started", signal = signal_name, "Shutting down JSON-RPC server.");
-        if let Err(err) = shutdown_handle.stop() {
-            warn!(error = %err, "Failed to stop JSON-RPC server cleanly");
-        }
-
-        // Stay live for a second signal and force-exit. Tokio's OS-level
-        // signal handler keeps intercepting SIGTERM/SIGINT even after the
-        // first one fires (tokio-rs/tokio#7905); if we let our Signal
-        // instances drop, a second Ctrl+C would be silently swallowed and
-        // a stuck graceful-shutdown could only be killed with SIGKILL.
-        // Re-await the already-registered handlers and exit non-zero on
-        // the second hit so an operator can always reclaim the process.
-        if recv_termination_signal(&mut sigterm, &mut sigint).await.is_none() {
-            return;
-        }
-        warn!(event = "force_exit", "Received second termination signal; forcing exit.");
-        std::process::exit(1);
-    });
+    // Make container teardown visible in logs and let in-flight proofs finish.
+    spawn_signal_bridge(server_handle.clone());
 
     server_handle.stopped().await;
     info!(event = "shutdown_complete", "JSON-RPC server stopped.");
